requirements.txt
uvicorn
fastapi
langchain
python-dotenv
openai
google-cloud-aiplatform >= 1.25.0
